lib/config_file.c lib/options.h lib/rfc959.c - fixed warnings about
2004-8-9 Brian Masney <masneyb@gftp.org>

        * lib/config_file.c lib/options.h lib/rfc959.c - fixed warnings about
        unitialized members in the structure

        * lib/pty.c (gftp_exec) - setup stderr when running the program

        * lib/sshv2.c src/uicommon/gftpui.h - fixed signed/unsigned integer
        comparisions

        * lib/sshv2.c - added sshv2_copy_param_options()

        * src/gtk/options_dialog.c - compile fix when compiling against GTK+
        1.2

+++ src/uicommon/gftpui.h
@@ -55,7 +55,7 @@ typedef enum _gftpui_common_request_type
 typedef struct _gftpui_common_methods
 {
   char *command;
-  int minlen;
+  size_t minlen;
   int (*func)(void *uidata, gftp_request * request,
               void *other_uidata, gftp_request * other_request,
               const char *command);
